% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_handling.R
\name{dataset}
\alias{dataset}
\title{The functionS in this file prepare the dataset for the modeling}
\usage{
dataset(
  varnum,
  setting = "No_Correlation",
  var = c("Mar", "No_Mar", "No_Var"),
  seed = 2,
  main_var = 10,
  var_effect = 0.5,
  correlation_var = 15,
  correlation_val = 5,
  high_dim = T,
  train_sample = 500
)
}
\description{
The functionS in this file prepare the dataset for the modeling
}
