% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fakename_gen.R
\name{fakename_gen}
\alias{fakename_gen}
\title{Create the Fakename and constant names in the model}
\usage{
fakename_gen(datafile, train_test = T, outcome_var = NA, num_outcome = 0)
}
\arguments{
\item{datafile}{dataframe or vector list to change the names}

\item{train_test}{to determine if the data contain separate train and test files}

\item{outcome_var}{to determine if the dataset has any outcome variable}

\item{num_outcome}{to detemrine the number of outcome varaibles. It takes variable name as input}
}
\value{
dataframe or vectorlist with changed names as well as list of realnames
}
\description{
This function takes a dataframe/vector list as an input and create the standard column and row names.
The column format is "X(numeric)" like "X1" and "X2". Rows are labelled as 1,2,3,4...
}
