% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDSI_parameters.R
\name{HDSI_formula_gen}
\alias{HDSI_formula_gen}
\title{Get the performance of HDSI model on any given dataset}
\usage{
HDSI_formula_gen(other_para = para)
}
\description{
the functions in this file provide support to HDSI_model.R file
}
\details{
#' @export
}
