% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDSI_model.R
\name{HDSI_model}
\alias{HDSI_model}
\title{Get the performance of HDSI model on any given dataset}
\usage{
HDSI_model(
  model = "all",
  inputdf = df,
  outvar = "y",
  seed = 1,
  bootstrap = T,
  effectsize = "large",
  k = 7,
  cint = 0.95,
  sd_level = 1,
  model_tech = "reg",
  para = HDSI_para_control(interactions = T, int_term = 2, intercept = T, out_type =
    "continuous", perf_metric = c("beta", "mp", "mp_beta")),
  covariate = c(1, "X8_"),
  min_max = c("min", "ci", "quartile")
)
}
\arguments{
\item{model}{determines the HDSI models that need to be run}

\item{inputdf}{is the dataframe containing the training and the test datasets}

\item{outvar}{is the outcome variable in the dataset.}

\item{para}{are the hyperparameters used to tweak the model settings.
It includes interactions, interaction level, number of latent factors for PLS, intercept term and datatype.}

\item{covariate}{takes the covariate which needs to be controlled during the feature selection process.
This is functional only for forward selection approach.}
}
\value{
The performance of the different models alongwith feature selection by the models
}
\description{
The functions takes a dataframe list as an input and evalaute the performance of different HDSI modeling techniques
like LASSO, ALASSO, RIDGE, ARIDGE, Forward, and Regression
}
