% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDSI_Tech.R
\name{mcoxph}
\alias{mcoxph}
\title{Get the performance of HDSI model on any given dataset}
\usage{
mcoxph(
  formula,
  data,
  weights,
  subset,
  na.action,
  init,
  control,
  ties = c("efron", "breslow", "exact"),
  singular.ok = TRUE,
  robust,
  model = FALSE,
  x = FALSE,
  y = TRUE,
  tt,
  method = ties,
  id,
  cluster,
  istate,
  statedata,
  ...
)
}
\description{
the functions in this file provide support to HDSI_model.R file
}
