% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostrap_gen.R
\name{p_est}
\alias{p_est}
\title{Prepare the bootstraps for any dataset}
\usage{
p_est(
  p,
  k = 3,
  rows = NA,
  interaction_numb = 2,
  effectsize = c("large", "medium", "small")
)
}
\arguments{
\item{p}{number of features}

\item{k}{number of features in a sample}

\item{rows}{number of rows of the dataframe}

\item{interaction_numb}{level of interactions considered}

\item{effectsize}{The expected difference between non-zero beta coefficient and zero value coefficient}

\item{feature_name}{is the list of features which will be sampled in each bootstrap}

\item{inputdf}{dataframe of the input data}

\item{type}{is the datatype which can be continuous or survival}

\item{seed_multiplier}{seed value to ensure result reproducibility}

\item{bootstrap}{to determine if the sampling is with or without replacement for rows}
}
\value{
vectorlist of all the bootsamples of rows and columns
}
\description{
The functions takes a dataframe/vector list as an input and perform the bootstrapping of rows
and sampling of the columns
}
