% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hyperparameter_optimisation.R
\name{quadraticRoots}
\alias{quadraticRoots}
\title{Optimise the hyperparameters}
\usage{
quadraticRoots(a, b, c)
}
\description{
Genetic Algorithm is applied to perform the hyperparameter optimisation.
}
